ex 2.17 updated

+++ part2/ex.2.15-2.18/src/App.js
@@ -56,11 +56,13 @@ const App = () => {
     setSearchField(event.target.value)
   }
   const handleDeletion = (event) => {
+    let person = persons.find(person => person.id === parseInt(event.target.id))
+    if (window.confirm(`Delete ${person.name} ?`)){
     AxiosCalls
     .deletePerson(event.target.id)
     .then(setPersons(
       persons.filter(person => person.id !== parseInt(event.target.id))))
- 
+    }
   }
   
   return (
